cast before multiplication to prevent potential overflow

+++ vector/v.vol.rst/user1.c
@@ -440,7 +440,8 @@ int OUTGR()
         G_fseek(Tmp_fd_z, 0L, 0);
 
         /* Read data in from temp file */
-        read_val = fread(data, sizeof(float), nsizr * nsizc * nsizl, Tmp_fd_z);
+        read_val =
+            fread(data, sizeof(float), (size_t)nsizr * nsizc * nsizl, Tmp_fd_z);
         if (read_val != (size_t)nsizr * nsizc * nsizl) {
             clean();
             G_fatal_error(_("Unable to read data from temp file"));
@@ -494,7 +495,8 @@ int OUTGR()
         G_fseek(Tmp_fd_dx, 0L, 0);
 
         /* Read data in from temp file */
-        read_val = fread(data, sizeof(float), nsizr * nsizc * nsizl, Tmp_fd_dx);
+        read_val = fread(data, sizeof(float), (size_t)nsizr * nsizc * nsizl,
+                         Tmp_fd_dx);
         if (read_val != (size_t)nsizr * nsizc * nsizl) {
             clean();
             G_fatal_error(_("Unable to read data from temp file"));
@@ -547,7 +549,8 @@ int OUTGR()
         G_fseek(Tmp_fd_dy, 0L, 0);
 
         /* Read data in from temp file */
-        read_val = fread(data, sizeof(float), nsizr * nsizc * nsizl, Tmp_fd_dy);
+        read_val = fread(data, sizeof(float), (size_t)nsizr * nsizc * nsizl,
+                         Tmp_fd_dy);
         if (read_val != (size_t)nsizr * nsizc * nsizl) {
             clean();
             G_fatal_error(_("Unable to read data from temp file"));
@@ -600,7 +603,8 @@ int OUTGR()
         G_fseek(Tmp_fd_dz, 0L, 0);
 
         /* Read data in from temp file */
-        read_val = fread(data, sizeof(float), nsizr * nsizc * nsizl, Tmp_fd_dz);
+        read_val = fread(data, sizeof(float), (size_t)nsizr * nsizc * nsizl,
+                         Tmp_fd_dz);
         if (read_val != (size_t)nsizr * nsizc * nsizl) {
             clean();
             G_fatal_error(_("Unable to read data from temp file"));
@@ -653,7 +657,8 @@ int OUTGR()
         G_fseek(Tmp_fd_xx, 0L, 0);
 
         /* Read data in from temp file */
-        read_val = fread(data, sizeof(float), nsizr * nsizc * nsizl, Tmp_fd_xx);
+        read_val = fread(data, sizeof(float), (size_t)nsizr * nsizc * nsizl,
+                         Tmp_fd_xx);
         if (read_val != (size_t)nsizr * nsizc * nsizl) {
             clean();
             G_fatal_error(_("Unable to read data from temp file"));
@@ -706,7 +711,8 @@ int OUTGR()
         G_fseek(Tmp_fd_yy, 0L, 0);
 
         /* Read data in from temp file */
-        read_val = fread(data, sizeof(float), nsizr * nsizc * nsizl, Tmp_fd_yy);
+        read_val = fread(data, sizeof(float), (size_t)nsizr * nsizc * nsizl,
+                         Tmp_fd_yy);
         if (read_val != (size_t)nsizr * nsizc * nsizl) {
             clean();
             G_fatal_error(_("Unable to read data from temp file"));
@@ -759,7 +765,8 @@ int OUTGR()
         G_fseek(Tmp_fd_xy, 0L, 0);
 
         /* Read data in from temp file */
-        read_val = fread(data, sizeof(float), nsizr * nsizc * nsizl, Tmp_fd_xy);
+        read_val = fread(data, sizeof(float), (size_t)nsizr * nsizc * nsizl,
+                         Tmp_fd_xy);
         if (read_val != (size_t)nsizr * nsizc * nsizl) {
             clean();
             G_fatal_error(_("Unable to read data from temp file"));
